Even more changes
- Tweaked several commands
- Merged /ticketsetup & /setmuterole into /config
- Fully removed maintenance mode
- Removed guildCreate event that registered commands
- Removed semi-guild command deploying
