+++ src/commands/fun/phcomment.js
@@ -18,13 +18,6 @@ module.exports = {
   async execute(interaction) {
     const text = interaction.options.getString("text");
 
-    //var responses = [
-      //"Friend, I'm gonna be honest with you. `*breaths in*` Your comment is complete ass.",
-      //"Friend, you're a genius. This comment is literally as perfect as you.",
-      //"Bestie, what the fuck is this?",
-      //"Friend, to be completely honest, this comment is why your mother considered putting you up for adoption.",
-  //];
-
     fetch(`https://nekobot.xyz/api/imagegen?type=phcomment&username=${interaction.user.username}&image=${interaction.user.avatarURL({ format: "png", size: 512 })}&text=${text}`)
     
     .then(function(result) { return result.json(); })
@@ -34,7 +27,7 @@ module.exports = {
             .setAuthor({name: "Here's your PH comment, you horny little shit."})
             .setImage(data.message)
             .setTimestamp()
-          return interaction.reply({ embeds: [phEmbed]} ) //content: responses[Math.floor(Math.random() * responses.length)],
+          return interaction.reply({ embeds: [phEmbed] });
     });
   },
 };
