+++ src/events/interaction/interactionCreate.js
@@ -7,15 +7,6 @@ module.exports = {
      * @param {Client} client 
      */
     async execute(interaction, client) {
-        if (client.maintenance && interaction.user.id != "292743562213457920") {
-            const Response = new MessageEmbed()
-                .setTitle("NOTICE")
-                .setDescription("Jinx is currently in emergency maintenance mode. [Join the support server for updates.](https://discord.gg/HwkDSs7X82)")
-                .setFooter({ text: "Take a break and have a coffee while we work." })
-                .setColor("DARK_PURPLE")
-            return interaction.reply({ embeds: [Response] })
-        }
-
         if (interaction.isCommand() || interaction.isContextMenu()) {
             const command = client.commands.get(interaction.commandName);
             if (!command) return interaction.reply({
